Feature: cached properties warning
 * Moved cached_property override to `utils.py`
 * Included a basic test to ensure a warning is called when accessed via
the __get__ method
