+++ ads/search.py
@@ -5,7 +5,6 @@ Interface to the adsws search api.
 import warnings
 import six
 import math
-from werkzeug.utils import cached_property as _cached_property
 import json
 
 from .config import SEARCH_URL
@@ -13,25 +12,7 @@ from .exceptions import SolrResponseParseError, APIResponseError
 from .base import BaseQuery, APIResponse
 from .metrics import MetricsQuery
 from .export import ExportQuery
-
-
-class cached_property(_cached_property):
-    """
-    Wrap cached property to print relevant warnings
-    """
-    def __init__(self, func, name=None, doc=None):
-        super(cached_property, self).__init__(func, name, doc)
-
-    def __get__(self, obj, type=None):
-        # One time warning that the user is using lazy loading
-        warnings.warn(
-            "You are lazy loading attributes via '{}', and so are "
-            "making multiple calls to the API. This will impact your overall "
-            "rate limits."
-            .format(self.func.__name__),
-            UserWarning,
-        )
-        return super(cached_property, self).__get__(obj, type)
+from .utils import cached_property
 
 
 class Article(object):
